difficolty over time and difficolty selector uodates

+++ MobileEndlessShooter/Assets/Scripts/EnemyScripts/EnemyData.cs
@@ -9,7 +9,7 @@ public class EnemyData : ScriptableObject
     [Header("Movement")] public float movementSpeed = 4f;
 
     [Header("Combat Stats")] public int health = 10;
-    public int damageToPlayer = 1;
+    public float damageToPlayer = 1;
 
     [Header("Scoring")] public int scoreValue = 100;
 }
